+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#299](https://github.com/dremio/dbt-dremio/pull/299) Enhance persist_docs macro to wrap model and column metadata (including descriptions, tags and tests) into a Markdown wiki for Dremio.
 - Refactored CI
 - Fixed tests for hooks and grants
+- Added Dremio Enterprise Catalog tests
 
 # dbt-dremio v1.9.0